feat: make landing page SSR-friendly for curl/crawler access
Convert LandingPage from a client component to a server component
so static content (title, description, CTAs, footer) renders in the
initial HTML response. Extract interactive AgentTabs into a separate
client component.

Amp-Thread-ID: https://ampcode.com/threads/T-019c3509-8b93-7775-bb21-a4ab19a5430e
